% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GearSpeedCheck.R
\name{GearSpeedCheck}
\alias{GearSpeedCheck}
\title{GearSpeedCheck}
\usage{
GearSpeedCheck(dataframe, timezone, maxSpeed, speedUnits, gearType)
}
\arguments{
\item{dataframe}{a dataframe with at least the following columns TIMESTAMP, LAT,
LON. TIMESTAMP should be formatted as yyyy-mm-dd HH:MM:SS. LAT and LON should
be in decimal degrees. Each data frame must have at least two rows of data.}

\item{timezone}{a character string indicating the timezone associated with
the specified timestamp. Appropriate values can be found using the command
OlsonNames()
The default is UTC}

\item{maxSpeed}{the maximum allowable speed. Defaults to 3 meters per second
(~ 8 knots) for all mobile gear.}

\item{speedUnits}{the units that the maximum speed is specified in. Defaults
to 'mps' (meters per second), other options include 'mph' (miles per hour)
and 'kts' (knots).}

\item{gearType}{a 2016 ISSCFG Gear Code or abbreviation (full list here: https://www.fao.org/3/bt988e/bt988e.pdf) There is no default value.}
}
\value{
Returns the original dataframe with an additional column called
"BadSpeed". The new column contains a TRUE/FALSE value describing whether the
points in the data occur at a bad (improbable) speed (TRUE)
}
\description{
This test is only applied to profiles coming from mobile gear. It controls
whether there are no erroneous locations provided. The speed of the vessels
are generated given the positions and times of the vessel. Vessel speed is
expected not to exceed 3 ms−1. Otherwise, it means either the positions or
times are bad data, or a vessel is sailing full speed rather than fishing.
This test is helpful for determining if there is an error in merging the
sensor and GPS data, often due to setting a sensor to a time zone other than
UTC.
}
\examples{
df=data.frame(
  LAT=c(42.16915,42.16916),
  LON=c(-66.92022,-66.92025),
  TIMESTAMP=c('2022-10-24 10:00:30','2022-10-24 10:05:30'),
  temperature=c(10,10.2),
  depth=c(100,101)
  )

GearSpeedCheck(
  dataframe=df,
  timezone='US/Eastern',
  maxSpeed=8,
  speedUnits='kts',
  gearType='OTB'
  )
}
