% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GearTypeCheck.R
\name{GearTypeCheck}
\alias{GearTypeCheck}
\title{GearTypeCheck
This function checks whether the distance traveled by the sensor during the
profile is appropriate for the type of gear deployed by the vessel. Distance
traveled is calculated between the first and last data locations.
Theoretically, fixed gear should not move very far from where it is deployed.
However, mobile gear should move a minimum distance (which will be fishery
dependent). For example, tows by draggers targeting groundfish may be
hundreds of meters while tows with hydraulic clam dredges may only be tens of
meters.}
\usage{
GearTypeCheck(startLon, startLat, endLon, endLat, gearType, buffer = NA)
}
\arguments{
\item{gearType}{a 2016 ISSCFG Gear Code or abbreviation (full list here: https://www.fao.org/3/bt988e/bt988e.pdf) There is no default value.}

\item{buffer}{a single numeric value representing the acceptable movement
distance (m) based on the  gear type. Eventually, each gear type should have
a programmed value that can be overridden by the user. For now, the user must
specify the value in most cases. It is important to note that for fixed
gears, the function will check if movement > buffer, and for mobile gears the
function will check if movement <= buffer.}

\item{dataframe}{a dataframe with at least the following columns TIMESTAMP,
LAT, LON. TIMESTAMP should be formatted as yyyy-mm-dd HH:MM:SS. LAT and LON
should be in decimal degrees. Each data frame must have at least two rows of data.}
}
\value{
Returns a TRUE/FALSE value with "TRUE" indicating that the gear has
moved in an unexpected way
}
\description{
GearTypeCheck
This function checks whether the distance traveled by the sensor during the
profile is appropriate for the type of gear deployed by the vessel. Distance
traveled is calculated between the first and last data locations.
Theoretically, fixed gear should not move very far from where it is deployed.
However, mobile gear should move a minimum distance (which will be fishery
dependent). For example, tows by draggers targeting groundfish may be
hundreds of meters while tows with hydraulic clam dredges may only be tens of
meters.
}
\examples{
GearTypeCheck(
     startLon=-70.92033,
     startLat=41.63875,
     endLon=-66.92022,
     endLat=42.16915,
     gearType='GNS'
     )
}
