% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRC_pressure.R
\name{GRC_pressure}
\alias{GRC_pressure}
\title{GRC_pressure}
\usage{
GRC_pressure(dataframe, column, make, model)
}
\arguments{
\item{dataframe}{a dataframe containing a column of pressure values reported
in decibars}

\item{column}{a character string indicating the column name containing the
pressure values}

\item{make}{The make of the sensor used to collect the data. Acceptable
options can be found using ListSensors()}

\item{model}{The model of the sensor used to collect the data. Acceptable
options can be found using ListSensors()}
}
\value{
Returns the original dataframe with an additional column called
"InvalidPressure". The new column contains a TRUE/FALSE value describing
whether the points in the data occur outside of the sensor's limits (TRUE)
}
\description{
This function checks whether the reported pressure value is within the range
of the sensor used.
}
\examples{
df=data.frame(
  LAT=c(42.16915,42.16916),
  LON=c(-66.92022,-66.92025),
  TIMESTAMP=c('2022-10-24 10:00:30','2022-10-24 10:05:30'),
  temperature=c(10,10.2),
  depth=c(199,202),
  do=c(9,9.5)
  )
GRC_pressure(
  dataframe=df,
  column='depth',
  make='ZebraTech',
  model='Moana-200'
  )
}
