% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRC_salinity.R
\name{GRC_salinity}
\alias{GRC_salinity}
\title{GRC_salinity}
\usage{
GRC_salinity(dataframe, column, make, model)
}
\arguments{
\item{dataframe}{a dataframe containing a column of dissolved salinity values
reported in practical salinity units}

\item{column}{a character string indicating the column name containing the
oxygen values}

\item{make}{The make of the sensor used to collect the data. Acceptable
options can be found using ListSensors()}

\item{model}{The model of the sensor used to collect the data. Acceptable
options can be found using ListSensors()}
}
\value{
Returns the original dataframe with an additional column called
"InvalidSalinity". The new column contains a TRUE/FALSE value describing
whether the points in the data occur outside of the sensor's limits (TRUE)
}
\description{
This function checks whether the reported salinity value is within the range
of the sensor used.
}
\examples{
df=data.frame(
  LAT=c(42.16915,42.16916),
  LON=c(-66.92022,-66.92025),
  TIMESTAMP=c('2022-10-24 10:00:30','2022-10-24 10:05:30'),
  temperature=c(10,10.2),
  depth=c(100,101),
  do=c(9,9.5),
  salinity=c(10,500)
  )
GRC_salinity(
  dataframe=df,
  column='salinity',
  make='NKE',
  model='Wisens CTDS'
  )
}
