% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarborCheck.R
\name{HarborCheck}
\alias{HarborCheck}
\title{HarborCheck}
\usage{
HarborCheck(homeport, lon, lat, buffer, uom = "meters")
}
\arguments{
\item{homeport}{a vector containing the longitude and latitude of the homeport. Longitude and latitude should be specified in decimal degrees in the order c(lon, lat). There is no default value.}

\item{lon}{a single numeric value representing the longitude where the observation was collected. Longitude should be specified in decimal degrees. There is no default value}

\item{lat}{a single numeric value representing the latitude where the observation was collected. Latitude should be specified in decimal degrees. There is no default value}

\item{buffer}{a single numeric value representing the size of the buffer (radius) around the homeport. There is no default value.}

\item{uom}{a character value representing the unit of measure that the buffer is specified in. Defaults to 'meters'. Other acceptable values include 'kilometers', 'nautical_miles', and 'miles'.}
}
\value{
Returns a TRUE/FALSE value as to whether the point specified by lat, lon is within the buffer around the point specified by homeport.
}
\description{
This function checks to see if data fall within a specified buffer zone
around a point of interest (typically the homeport of a vessel).
}
\examples{
HarborCheck(
     homeport=c(-70.92033,41.63875),
     lon=-66.92022,
     lat=42.16915,
     buffer=250,
     uom='nautical_miles'
     )
}
