% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpossibleLocationCheck.R
\name{ImpossibleLocationCheck}
\alias{ImpossibleLocationCheck}
\title{ImpossibleLocationCheck}
\usage{
ImpossibleLocationCheck(dataframe)
}
\arguments{
\item{dataframe}{a dataframe containing at least the columns LATITUDE and LONGITUDE}
}
\value{
Returns the original dataframe with an extra column called ImpossibleLocation in which TRUE indicates bad data.
}
\description{
This function checks to see if the location of the data collection is out of
bounds
}
\examples{
ImpossibleLocationCheck(
  dataframe=df
  )
}
