% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListSensors.R
\name{ListSensors}
\alias{ListSensors}
\title{ListSensors}
\usage{
ListSensors(make = NA, model = NA)
}
\arguments{
\item{make}{filter only sensors of this make. Defaults to NA.}

\item{model}{filter only sensors of this model. Defaults to NA.}
}
\value{
a vector of possible Make/Model combinations for use in this package
}
\description{
ListSensors
}
\examples{
ListSensors()
}
