% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OnLandCheck.R
\name{OnLandCheck}
\alias{OnLandCheck}
\title{OnLandCheck}
\usage{
OnLandCheck(data)
}
\arguments{
\item{dataframe}{a dataframe containing at minimum a column for latitude and a
column for longitude. These columns should be named LAT and LON. There is no
default value.}
}
\value{
Returns the original dataframe with an additional column called
"OnLand". The new column contains a TRUE/FALSE value describing whether the
points in the data are on land or not. TRUE values indicate data are on land.
}
\description{
This function checks to see if the data fall within areas defined as land or
water based on the high resolution GeoJSON available from the World Bank at
https://datacatalog.worldbank.org/search/dataset/0038272
}
\examples{
df=data.frame(
  LAT=42.16915,
  LON=-66.92022,
  temperature=10,
  depth=100
  )

OnLandCheck(data=df)
}
