% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateOfChangeCheck.R
\name{RateOfChangeCheck}
\alias{RateOfChangeCheck}
\title{RateofChangeCheck}
\usage{
RateOfChangeCheck(column, threshold = 3, dataframe)
}
\arguments{
\item{column}{The column name to apply the test to. There is no default value.}

\item{threshold}{The threshold (in SD) of change that is considered
acceptable. The default value is 3.}

\item{dataframe}{a dataframe containing at least the column TIMESTAMP in the
format yyyy-mm-dd HH:MM:SS and one parameter of interest. There is no default
value.}
}
\value{
Returns the original dataframe with an additional column called
"FailRateCheck". The new column contains a TRUE/FALSE value describing
whether the points fail the rate check (bad data = TRUE).
}
\description{
This test is applied per segment (Up-Down-Bottom), and
inspects the segments on a rate of change exceeding a threshold defined by
the operator. In this case the thresholds are based on the IOOS examples
(U.S. Integrated Ocean Observing System, 2020), where the rate of change
between measurement Tn-1 and Tn must be less than three standard deviations
(3*SD). The SD of the T time series is computed over the full segment.
}
\examples{
RateofChangeCheck(
     column="TEMPERATURE",
     threshold=3,
     dataframe=df
     )
}
