% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RollingAvgSpikeCheck.R
\name{RollingAvgSpikeCheck}
\alias{RollingAvgSpikeCheck}
\title{RollingAvgSpikeCheck}
\usage{
RollingAvgSpikeCheck(
  dataframe,
  column,
  time_column,
  rolling_avg = 30,
  threshold = 1
)
}
\arguments{
\item{dataframe}{a dataframe containing timestamps and observations}

\item{column}{a character string indicating the column name to be checked
for spikes}

\item{time_column}{a character string indicating the column name containing
the timestamp values. Timestamps should be in yyyy-mm-dd hh:mm:ss format. No default value.}

\item{rolling_avg}{an integer indicating the number of points to include in
the rolling average (defaults to 30)}

\item{threshold}{a number indicating the standard deviation threshold to
trigger a QAQC flag (defaults to 1)}
}
\value{
Returns the original dataframe with an additional column called
"FailRASpikeCheck". The new column contains a TRUE/FALSE value describing
whether the points fail the rate check (bad data = TRUE).
}
\description{
The RollingAvgSpikeCheck function checks whether each observation exceeds or
is lower than the rolling average by some number of standard deviations
(default threshold = 1 SD).
}
\examples{
RollingAvgSpikeCheck(
     column="TEMPERATURE",
     threshold=3,
     dataframe=df,
     time_column='TIMESTAMP',
     )
}
