% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeCheck_Salinity.R
\name{SpikeCheck_Salinity}
\alias{SpikeCheck_Salinity}
\title{SpikeCheck_Salinity}
\usage{
SpikeCheck_Salinity(dataframe, salinity_column, depth_column, time_column)
}
\arguments{
\item{dataframe}{a dataframe containing at least the following information:
\enumerate{
\item a column of timestamps in the format 'yyyy-mm-dd HH:MM:SS'
\item a column of salinity in PSU
\item a column of depths in dbars
}}

\item{salinity_column}{a character string indicating the column name containing
the salinity values}

\item{depth_column}{a character string indicating the column name containing
the depth values}

\item{time_column}{a character string indicating the column name containing
the timestamp values}
}
\value{
Returns the original dataframe with an additional column called
"SalinitySpike". The new column contains a TRUE/FALSE value describing
whether the points in the data are identified as spikes (TRUE)
}
\description{
The SpikeCheck_Semp function checks whether there is a significant difference
between sequential salinity measurements, by comparing a measurement to
the adjacent ones. The dataframe must contain at least three lines of data.
}
\examples{
df=data.frame(
  LAT=c(42.16915,42.16916,42.16917),
  LON=c(-66.92022,-66.92025,-66.92026),
  TIMESTAMP=c('2022-10-24 10:00:30','2022-10-24 10:05:30','2022-10-24 10:10:30'),
  temperature=c(10,50,10.2),
  depth=c(100,101,100),
  do=c(9,109.5,9),
  salinity=c(35,38,35.1)
  )
SpikeCheck_Salinity(
  dataframe=df,
  salinity_column='salinity',
  depth_column='depth',
  time_column='TIMESTAMP'
  )
}
