% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpikeCheck_Temp.R
\name{SpikeCheck_Temp}
\alias{SpikeCheck_Temp}
\title{SpikeCheck_Temp}
\usage{
SpikeCheck_Temp(dataframe, temp_column, time_column)
}
\arguments{
\item{dataframe}{a dataframe containing at least the following information:
\enumerate{
\item a column of timestamps in the format 'yyyy-mm-dd HH:MM:SS'
\item a column of temperatures in Celsius
\item a column of depths in dbars
}}

\item{temp_column}{a character string indicating the column name containing
the temperature values}

\item{time_column}{a character string indicating the column name containing
the timestamp values}

\item{depth_column}{a character string indicating the column name containing
the depth values}
}
\value{
Returns the original dataframe with an additional column called
"TempSpike". The new column contains a TRUE/FALSE value describing
whether the points in the data are identified as spikes (TRUE)
}
\description{
The SpikeCheck_Temp function checks whether there is a significant difference
between sequential temperature measurements, by comparing a measurement to
the adjacent ones. The dataframe must contain at least three lines of data.
}
\examples{
df=data.frame(
  LAT=c(42.16915,42.16916,42.16917),
  LON=c(-66.92022,-66.92025,-66.92026),
  TIMESTAMP=c('2022-10-24 10:00:30','2022-10-24 10:05:30','2022-10-24 10:10:30'),
  temperature=c(10,50,10.2),
  depth=c(100,101,100),
  do=c(9,109.5,9)
  )
SpikeCheck_Temp(
  dataframe=df,
  temp_column='temperature',
  depth_column='depth',
  time_column='TIMESTAMP'
  )
}
