% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnlikelyTempCheck.R
\name{UnlikelyTempCheck}
\alias{UnlikelyTempCheck}
\title{UnlikelyTempCheck}
\usage{
UnlikelyTempCheck(dataframe, tempcol, min_temp = 0, max_temp = 30)
}
\arguments{
\item{dataframe}{a dataframe containing a column of temperatures}

\item{tempcol}{a character string indicating which column name stores the time}

\item{min_temp}{an integer indicating the lowest acceptable temperature (defaults to 0)}

\item{max_temp}{an integer indicating the highest acceptable temperature (defaults to 30)}
}
\value{
Returns a dataframe containing a TRUE/FALSE value as to whether the temperature is out of bounds.
TRUE indicates bad data.
}
\description{
This function checks to see if the temperature records are wildly out of
bounds
}
\examples{
 df=data.frame(
  LAT=c(42.16915,42.16916),
  LON=c(-66.92022,-66.92025),
  TIMESTAMP=c('2022-10-24 10:00:30','2022-10-24 10:05:30'),
  temperature=c(10,10.2),
  depth=c(100,101),
  do=c(9,109.5)
  )
UnlikelyTempCheck(
  dataframe=df,
  tempcol="temperature",
  )
}
