% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedRegionCheck.R
\name{ExpectedRegionCheck}
\alias{ExpectedRegionCheck}
\title{ExpectedRegionCheck}
\usage{
ExpectedRegionCheck(region, dataframe)
}
\arguments{
\item{region}{the region the vessel is expected to be operating in. Can be
one of the following: 'Alaska','North_Atlantic','Greenland','New_Zealand',
'North_Sea_Baltic'. There is no default value.}

\item{dataframe}{a dataframe containing at least the columns LATITUDE and LONGITUDE,
where LATITUDE is latitude in decimal degrees and LONGITUDE is longitude in decimal
degrees. There is no default value.}
}
\value{
Returns the original dataframe with an additional column called
"OutRegion". The new column contains a TRUE/FALSE value describing whether
the points in the data occur outside of the specified region (TRUE)
}
\description{
This function checks to see if the data fall within the expected
operating region of a vessel.
}
\examples{
ExpectedRegionCheck(
     region='North_Atlantic',
     dataframe=df
     )
}
