% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarborCheck.R
\name{HarborCheck}
\alias{HarborCheck}
\title{HarborCheck}
\usage{
HarborCheck(homeport, dataframe, buffer, uom = "meters")
}
\arguments{
\item{homeport}{a vector containing the longitude and latitude of the
homeport. Longitude and latitude should be specified in decimal degrees in
the order c(lon, lat). There is no default value.}

\item{dataframe}{a dataframe with at least the following columns LAT and LON.
LAT and LON should be in decimal degrees. There is no default.}

\item{buffer}{a single numeric value representing the size of the buffer
(radius) around the homeport. There is no default value.}

\item{uom}{a character value representing the unit of measure that the buffer
is specified in. Defaults to 'meters'. Other acceptable values include
'kilometers', 'nautical_miles', and 'miles'.}
}
\value{
Returns the original dataframe with two additional columns called
"dist_from_port" and "inPort".
"dist_from_port" contains a numeric value representing the straight line
distance from the port to the data point in uom specified by the user.
"inPort" contains a TRUE/FALSE value describing whether the points in the
data occur inside the user-defined harbor buffer (TRUE) or not.
}
\description{
This function checks to see if data fall within a specified buffer zone
around a point of interest (typically the homeport of a vessel).
}
\examples{
df=data.frame(
  LAT=c(42.16915,42.16916),
  LON=c(-66.92022,-66.92025),
  TIMESTAMP=c('2022-10-24 10:00:30','2022-10-24 10:05:30'),
  temperature=c(10,10.2),
  depth=c(100,101)
  )
HarborCheck(
     homeport=c(-70.92033,41.63875),
     dataframe=df,
     buffer=1,
     uom='nautical_miles'
     )
}
