% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpossibleLocationCheck.R
\name{ImpossibleLocationCheck}
\alias{ImpossibleLocationCheck}
\title{ImpossibleLocationCheck}
\usage{
ImpossibleLocationCheck(lat, lon)
}
\arguments{
\item{lat}{latitude of the data collection. There is no default value.}

\item{lon}{longitude of the data collection. There is no default value.}
}
\value{
Returns a TRUE/FALSE value as to whether location is out of bounds.
TRUE indicates bad data.
}
\description{
This function checks to see if the location of the data collection is out of
bounds
}
\examples{
ImpossibleLocationCheck(
  lat=42.16915,
  lon=-66.92022
  )
}
